Refactor: Seperate function
