+++ lib/constant/state.ts
@@ -3,7 +3,6 @@ const STATE = {
   IS_ATTR_NAME: 'isAttrName',
   IS_ATTR_VALUE: 'isAttrValue',
   IS_VALUE: 'isValue',
-  IS_PARTIAL_TEMPLATE: 'isPartialTemplate',
 } as const;
 
 export default STATE;
